add gadgets/mri_core/GenericReconGadget
