+++ gadgets/mri_core/GenericCartesianGrappaReconGadget.h
@@ -5,28 +5,7 @@
 
 #pragma once
 
-#include <complex>
-#include "gadgetron_mricore_export.h"
-#include "Gadget.h"
-#include "hoNDArray.h"
-#include "ismrmrd/ismrmrd.h"
-#include "ismrmrd/xml.h"
-#include "ismrmrd/meta.h"
-#include "GadgetronTimer.h"
-
-#include "hoNDArray_utils.h"
-
-//#include "gtPlusIOAnalyze.h"
-
-#include "GadgetStreamController.h"
-
-#include "hoNDArray_utils.h"
-#include "hoNDArray_elemwise.h"
-#include "hoNDFFT.h"
-
-#include "mri_core_data.h"
-#include "mri_core_utility.h"
-#include "mri_core_coil_map_estimation.h"
+#include "GenericReconGadget.h"
 
 namespace Gadgetron {
 
@@ -73,12 +52,12 @@ namespace Gadgetron {
 
 namespace Gadgetron {
 
-    class EXPORTGADGETSMRICORE GenericCartesianGrappaReconGadget : public Gadget1<IsmrmrdReconData>
+    class EXPORTGADGETSMRICORE GenericCartesianGrappaReconGadget : public GenericReconGadget
     {
     public:
         GADGET_DECLARE(GenericCartesianGrappaReconGadget);
 
-        typedef Gadget1<IsmrmrdReconData> BaseClass;
+        typedef GenericReconGadget BaseClass;
         typedef Gadgetron::GenericCartesianGrappaReconObj< std::complex<float> > ReconObjType;
 
         GenericCartesianGrappaReconGadget();
@@ -88,26 +67,10 @@ namespace Gadgetron {
         /// parameters to control the reconstruction
         /// ------------------------------------------------------------------------------------
 
-        /// image series
-        GADGET_PROPERTY(image_series, int, "Image series number", 0);
-
-        /// ------------------------------------------------------------------------------------
-        /// debug and timing
-        GADGET_PROPERTY(verbose, bool, "Whether to print more information", false);
-        GADGET_PROPERTY(debug_folder, std::string, "If set, the debug output will be written out", "");
-        GADGET_PROPERTY(perform_timing, bool, "Whether to perform timing on some computational steps", false);
-
         /// ------------------------------------------------------------------------------------
         /// image sending
         GADGET_PROPERTY(send_out_gfactor, bool, "Whether to send out gfactor map", false);
         GADGET_PROPERTY(send_out_snr_map, bool, "Whether to send out SNR map", false);
-
-        /// ------------------------------------------------------------------------------------
-        /// ref preparation
-        /// whether to average all N for ref generation
-        GADGET_PROPERTY(average_all_ref_N, bool, "Whether to average all N for ref generation", true);
-        /// whether to average all S for ref generation
-        GADGET_PROPERTY(average_all_ref_S, bool, "Whether to average all S for ref generation", false);
 
         /// ------------------------------------------------------------------------------------
         /// Grappa parameters
@@ -120,58 +83,10 @@ namespace Gadgetron {
     protected:
 
         // --------------------------------------------------
-        // variables for protocol
-        // --------------------------------------------------
-
-        // number of encoding spaces in the protocol
-        size_t num_encoding_spaces_;
-
-        // for every encoding space
-
-        // acceleration factor for E1 and E2
-        std::vector<double> acceFactorE1_;
-        std::vector<double> acceFactorE2_;
-
-        // calibration mode
-        std::vector<Gadgetron::ismrmrdCALIBMODE> calib_mode_;
-
-        // encoding space limits
-        std::vector<ISMRMRD::EncodingCounters> meas_max_idx_;
-
-        // --------------------------------------------------
         // variable for recon
         // --------------------------------------------------
         // record the recon kernel, coil maps etc. for every encoding space
         std::vector< ReconObjType > recon_obj_;
-
-        // number of times the process function is called
-        size_t process_called_times_;
-
-        /// buffers used during recon
-        hoNDArray< std::complex<float> > complex_im_recon_buf_;
-        hoNDArray< std::complex<float> > data_recon_buf_;
-
-        // filter used for ref coil map 
-        hoNDArray< std::complex<float> > filter_RO_ref_coi_map_;
-        hoNDArray< std::complex<float> > filter_E1_ref_coi_map_;
-        hoNDArray< std::complex<float> > filter_E2_ref_coi_map_;
-
-        // --------------------------------------------------
-        // variables for debug and timing
-        // --------------------------------------------------
-
-        // in verbose mode, more info is printed out
-        bool verbose_;
-
-        //// debug folder
-        //std::string debug_folder_full_path_;
-
-        // clock for timing
-        Gadgetron::GadgetronTimer gt_timer_local_;
-        Gadgetron::GadgetronTimer gt_timer_;
-
-        //// exporter
-        //Gadgetron::gtPlus::gtPlusIOAnalyze gt_exporter_;
 
         // --------------------------------------------------
         // gadget functions
@@ -184,28 +99,10 @@ namespace Gadgetron {
         // recon step functions
         // --------------------------------------------------
 
-        // make the ref data for coil map estimation
-        virtual void make_ref_coil_map(IsmrmrdDataBuffered& ref_,std::vector<size_t> recon_dims, ReconObjType& recon_obj, size_t encoding);
-
-        // estimate coil map
-        virtual void perform_coil_map_estimation(IsmrmrdReconBit& recon_bit, ReconObjType& recon_obj, size_t encoding);
-
         // calibration, if only one dst channel is prescribed, the GrappaOne is used
         virtual void perform_calib(IsmrmrdReconBit& recon_bit, ReconObjType& recon_obj, size_t encoding);
 
         // unwrapping or coil combination
         virtual void perform_unwrapping(IsmrmrdReconBit& recon_bit, ReconObjType& recon_obj, size_t encoding);
-
-        // compute image header
-        virtual void compute_image_header(IsmrmrdReconBit& recon_bit, ReconObjType& recon_obj, size_t encoding);
-
-        // send out the recon results
-        virtual int send_out_image_array(IsmrmrdReconBit& recon_bit, IsmrmrdImageArray& res, size_t encoding, int series_num, const std::string& data_role);
-
-        // --------------------------------------------------
-        // utility functions
-        // --------------------------------------------------
-        // compute image number
-        virtual size_t compute_image_number(ISMRMRD::ImageHeader& imheader, size_t encoding = 0, size_t CHA = 1, size_t cha = 0, size_t E2 = 1);
     };
 }
